requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Data Validation & Env
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
aiohttp==3.9.0
httpx==0.26.0

# Database
asyncpg==0.29.0
SQLAlchemy[asyncio]==2.0.25
alembic==1.13.1
redis==5.0.1

# Vector Databases (для RAG)
chromadb==0.4.18
qdrant-client==1.7.0
pgvector==0.2.4

# AI/ML
transformers==4.37.0
torch==2.1.0
numpy==1.26.0
scikit-learn==1.4.0
sentence-transformers==2.2.2
openai==1.12.0

# Task Queue
celery==5.3.4

# Security
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Rate Limiting
slowapi==0.1.8

# Integrations
PyGithub==2.1.0

# Media Processing
Pillow==10.2.0
librosa==0.10.1
soundfile==0.12.1

# Monitoring & Logging
sentry-sdk==1.40.0
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Production Server
gunicorn==21.2.0
